// frontend/src/app/shared/enums/node-sub-type.enum.ts
export enum NodeSubTypeEnum {
    Root = 'Root',
    Folder = 'Folder',
    GroupNode = 'Group',
    StandardGroup = 'StandardGroup',
    Project = 'Project',
    Module = 'Module',
    Widget = 'Widget',
    View = 'View',
    Ddb = 'Ddb',
    ConcatText = 'ConcatText',
    Teleprompter = 'Teleprompter',
    List = 'List',
    MetaText = 'MetaText',
    DisplayText = 'DisplayText',
    SdsCommand = 'SdsCommand',
    ConcatString = 'ConcatString',
    SdsText = 'SdsText',
    SdsPrompt = 'SdsPrompt',
    SpeechCommand = 'SpeechCommand',
    SpeechPrompt = 'SpeechPrompt',
}